requirements.txt
<<<<<<< HEAD
streamlit
pandas
langchain

=======
accelerate==0.24.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.5.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bentoml==1.1.8
bitsandbytes==0.41.1
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
ci-info==0.3.0
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.0.0
coloredlogs==15.0.1
configobj==5.0.8
configparser==6.0.0
contextlib2==21.6.0
contractions==0.1.73
dataclasses-json==0.5.14
datasets==2.14.6
deepmerge==1.1.0
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
emailmessage==0.2.1
etelemetry==0.3.0
fastapi==0.103.1
fastcore==1.5.29
filelock==3.12.3
filetype==1.2.0
fitz==0.0.1.dev2
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.10.0
future==0.18.3
ghapi==1.0.4
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.295
langsmith==0.0.38
looseversion==1.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numexpr==2.8.6
numpy==1.25.2
openai==0.28.0
openllm==0.3.14
openllm-client==0.3.14
openllm-core==0.3.14
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
optimum==1.14.0
optional-django==0.3.0
orjson==3.9.10
packaging==23.1
pandas==2.1.0
param==1.13.0
pathlib==1.0.1
pathspec==0.11.2
Pillow==10.0.0
pip-requirements-parser==32.0.1
pip-tools==7.3.0
prometheus-client==0.18.0
protobuf==4.21.12
prov==2.0.0
psutil==5.9.6
pyahocorasick==2.0.0
pyarrow==14.0.0
pydantic==1.10.13
pydantic_core==2.6.3
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.16.1
pynvml==11.5.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pytesseract==0.3.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
pyzmq==25.1.1
rdflib==7.0.0
react==4.3.0
referencing==0.32.0
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.13.2
safetensors==0.4.0
schema==0.7.5
scipy==1.11.2
sentencepiece==0.1.99
simple-di==0.1.5
simplejson==3.19.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
-e git+https://github.com/Stability-AI/stability-sdk@59480881366e9dabdf60b47e4d2328d4b5c9ddc0#egg=stability_sdk
starlette==0.27.0
stopwords==1.0.0
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
textsearch==0.0.24
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traits==6.3.2
transformers==4.35.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.4
uvicorn==0.23.2
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.9
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
>>>>>>> 9da7614 (test)
